Removed c++17 dependency  (inline variables)
