+++ src/TimerModules/RTC/RTC.cpp
@@ -1,0 +1,30 @@
+#include "RTC.h"
+#include "RtcChannel.h"
+
+namespace TeensyTimerTool::RTC_t
+{
+    namespace // private
+    {
+        callback_t callback = nullptr;
+        RtcChannel channel(&callback);
+        bool initialized = false;
+
+        void isr()
+        {
+            SNVS_HPSR |= 0b11; // reset interrupt flag
+            if (callback != nullptr) callback();
+            asm volatile("DSB");
+        }
+    }
+
+    ITimerChannel *getTimer()
+    {
+        if (!initialized)
+        {
+            attachInterruptVector(IRQ_SNVS_IRQ, isr);
+            NVIC_ENABLE_IRQ(IRQ_SNVS_IRQ);
+            initialized = true;
+        }
+        return &channel;
+    }
+} // namespace TeensyTimerTool::RTC_t
